Inline DynamicProperties into Qualifiers
It's now only used in this one place, so just inline it rather than
including it.
